Implement ProgramOwnedTree and update payload for Merkle proofs
* Implemented the `ProgramOwnedTree` rule which was previously returned
`NotImplemented`.
* Moved Merkle proof calculation to util function since its used by two
rules now.
* Modifying `PubkeyTreeMatch` rule to have two fields, the same way
we solved it for `PDAMatch`.  This also solves some ambiguity around
how this rule would be used by a program such as token-metadata.
* This will allow us to follow a pattern where the user specifies the
Merkle proof, and token-metadata adds the destination Pubkey.
* Changed `LeafInfo` type to `ProofInfo` becuase leaf node is now
just specified as a Pubkey (similar to `PDAMatch`).
* Adding doc comments for all rules specifying exactly what is needed
in the payload and/or additional accounts when `Validate` is called.
* Add test for 'ProgramOwnedTree'.
* Regenerate IDL/SDK.