Replace payer not signer test with one that makes more sense
The previous test just was trying to do unsigned transactions
so the payer wasn't a signer and I don't think it could just run
the transaction for free.

The new test expects an "other_payer" to be present and is expected
to panic if only the banks client context payer signs.  I think this
is a more realistic situation.

I wanted to try to validate the processor check on if the payer is a
signer, but practically I do not think I can do that when using
Shank and the instruction builder.