Update IsWallet rule and basic royalty enforcement RuleSet in test
* Update doc comment for IsWallet rule to contain details about its
implementation and required payload value.
* Update IsWallet rule to check for System Program ownership in
addition to the existing commented out on-curve check.
* Update basic royalty enforcement RuleSet in test to allow
wallet-to-wallet.  Currently it will return failure because it is
not implemented but it is OR'd with the other conditions.
* Update the way errors are returned from the `Any` rule.
Essentially, return the error code from the last rule that failed,
but don't overwrite an existing last rule error code with
`NotImplemented`, otherwise pretty much every royalty enforcement
failure would return `NotImplemented` because it's last in the
RuleSet, and the RuleSet is ordered based on computation required.