Updating program owned tree test to use a nonzero data account
Added a test utility to generate a merkle tree based on input
leaf, so that I could pass in a known program to be the owner and
have it generate a proof.  With the previous hardcoded Merkle tree,
I was not easily able to use the leaf as an owner for an account
that had nonzero data (since I could not assign the leaf as owner
and then add data).

Also double check data length and ownership during the test using
banks client to retrieve account info.